requirements.txt
flask==3.0.0
werkzeug==3.0.0
python-dotenv==1.0.0
annotated-types-0.7.0 
distro-1.9.0
jiter-0.10.0
openai-1.102.0
pydantic-2.11.7
pydantic-core-2.33.2
tqdm-4.67.1
typing-inspection-0.4.1 